fix: create independend router tests with logic that separated the execution of the comand

+++ order/model/new_order.go
@@ -3,6 +3,6 @@ package model
 import "mc-burger-orders/item"
 
 type NewOrder struct {
-	CustomerId int16       `json:"customerId" binding:"required"`
+	CustomerId int         `json:"customerId" binding:"required"`
 	Items      []item.Item `json:"items" binding:"required,gt=0"`
 }
